phreaknet-inward-local: fix broken network operator routing
from-phreaknet-operator went to phreaknet-inward-semipublic, rather than
phreaknet-inward, which was intentional because phreaknet-inward includes
phreaknet-inward-nonpublic, which should not be accessible from any node
besides the operator services tandem. As a result, access to phreaknet-inward-local
was not possible, which broke routing to 11X codes.

The initial idea was that these codes should also be effectively "non public", but
to fix this, 11X is now merged into semipublic to fix the broken routing.