phreaknet.conf: Update CDR ticketing for res_phreaknet
Prefer to use the PHREAKNET function to set up a CDR hook,
rather than the legacy hangup handler in phreaknet-aux.conf.

In the dialplan snippet, it would really be better to use app_if.
However, this was merged only very recently, too recently to
expect any systems not installed using PhreakScript to have it.
As such, we use the clunkier but more compatible GotoIf syntax
for the time being here. In the future, it can be converted.